backend/requirements.txt
flask
flask-cors
flask-mysqldb
bcrypt
python-dotenv
anthropic
PyJWT
groq
google-genai
